🔨 Refactor: remove unused code & change client expiration time
🔨 refactor(BukkitClientModule.java): remove unused code
🔨 refactor(PlayerLeftListener.java): change client expiration time to 10 seconds

The unused code in the BukkitClientModule.java file has been removed. The code was responsible for loading clients for players, but it was not being used. The expiration time for clients in the PlayerLeftListener.java file has been changed from 5 minutes to 10 seconds. This change was made to reduce the time it takes for clients to expire and to improve performance.