🎉✨ Add: WarpCreatedEvent & WarpDeletedEvent
🎉 feat(WarpCommand.java): add WarpCreatedEvent & WarpDeletedEvent
✨ feat(WarpCreatedEvent.java): add event for when a warp is created
✨ feat(WarpDeletedEvent.java): add event for when a warp is deleted

The WarpCreatedEvent and WarpDeletedEvent classes have been added to the project. These events are triggered when a warp is created or deleted, respectively. The WarpCommand class has been updated to trigger these events when a warp is created or deleted. This allows other parts of the codebase to listen for these events and perform additional actions when a warp is created or deleted.